unsquashfs: Add --kind parsing of other squashfs formats
See: #181
